component classes and styles caching
