+++ src/textual/widgets/_option_list.py
@@ -1,124 +1,104 @@
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, ClassVar, Iterable, NamedTuple
+from dataclasses import dataclass, field
+from typing import TYPE_CHECKING, ClassVar, Iterable, Sequence, cast
 
 import rich.repr
-from rich.console import RenderableType
-from rich.measure import Measurement
-from rich.rule import Rule
-from rich.style import Style
+from rich.segment import Segment
 
 from textual import _widget_navigation, events
-from textual._widget_navigation import Direction
+from textual._loop import loop_last
 from textual.binding import Binding, BindingType
 from textual.cache import LRUCache
+from textual.css.styles import RulesMap
 from textual.geometry import Region, Size
 from textual.message import Message
 from textual.reactive import reactive
 from textual.scroll_view import ScrollView
 from textual.strip import Strip
-from textual.visual import Padding, Visual, visualize
+from textual.style import Style
+from textual.visual import Padding, Visual, VisualType, visualize
 
 if TYPE_CHECKING:
     from typing_extensions import Self, TypeAlias
 
-    from textual.app import RenderResult
+
+OptionListContent: TypeAlias = "Option | VisualType | None"
+"""Types accepted in OptionList constructor and [add_options()][textual.widgets.OptionList.ads_options]."""
 
 
-class DuplicateID(Exception):
+class OptionListError(Exception):
+    """An error occurred in the option list."""
+
+
+class DuplicateID(OptionListError):
     """Raised if a duplicate ID is used when adding options to an option list."""
 
 
-class OptionDoesNotExist(Exception):
+class OptionDoesNotExist(OptionListError):
     """Raised when a request has been made for an option that doesn't exist."""
-
-
-class Separator:
-    """Class used to add a separator to an [OptionList][textual.widgets.OptionList]."""
 
 
 @rich.repr.auto
 class Option:
-    """Class that holds the details of an individual option."""
+    """This class holds details of options in the list."""
 
     def __init__(
-        self, prompt: RenderResult, id: str | None = None, disabled: bool = False
+        self, prompt: VisualType, id: str | None = None, disabled: bool = False
     ) -> None:
         """Initialise the option.
 
         Args:
-            prompt: The prompt for the option.
-            id: The optional ID for the option.
-            disabled: The initial enabled/disabled state. Enabled by default.
+            prompt: The prompt (text displayed) for the option.
+            id: An option ID for the option.
+            disabled: Disable the option (will be shown grayed out, and will not be selectable).
+
         """
         self._prompt = prompt
+        self._visual: Visual | None = None
         self._id = id
-        self.disabled = disabled
+        self._disabled = disabled
+        self._divider = False
 
     @property
-    def prompt(self) -> RenderResult:
-        """The prompt for the option."""
-        return self._prompt
-
-    def set_prompt(self, prompt: RenderResult) -> None:
-        """Set the prompt for the option.
-
-        Args:
-            prompt: The new prompt for the option.
-        """
-        self._prompt = prompt
-
-    def visualize(self) -> object:
+    def prompt(self) -> VisualType:
+        """The original prompt."""
         return self._prompt
 
     @property
     def id(self) -> str | None:
-        """The optional ID for the option."""
+        """Optional ID for the option."""
         return self._id
 
+    @property
+    def disabled(self) -> bool:
+        return self._disabled
+
+    def __hash__(self) -> int:
+        return id(self)
+
     def __rich_repr__(self) -> rich.repr.Result:
-        yield "prompt", self.prompt
-        yield "id", self.id, None
+        yield self._prompt
+        yield "id", self._id, None
         yield "disabled", self.disabled, False
-
-    def __rich__(self) -> RenderResult:
-        return self._prompt
+        yield "_divider", self._divider, False
 
 
-class OptionLineSpan(NamedTuple):
-    """Class that holds the line span information for an option.
+@dataclass
+class _LineCache:
+    """Cached line information."""
 
-    An [Option][textual.widgets.option_list.Option] can have a prompt that
-    spans multiple lines. Also, there's no requirement that every option in
-    an option list has the same span information. So this structure is used
-    to track the line that an option starts on, and how many lines it
-    contains.
-    """
+    lines: list[tuple[int, int]] = field(default_factory=list)
+    heights: dict[int, int] = field(default_factory=dict)
+    index_to_line: dict[int, int] = field(default_factory=dict)
 
-    first: int
-    """The line position for the start of the option.."""
-    line_count: int
-    """The count of lines that make up the option."""
-
-
-OptionListContent: TypeAlias = "Option | Separator"
-"""The type of an item of content in the option list.
-
-This type represents all of the types that will be found in the list of
-content of the option list after it has been processed for addition.
-"""
-
-NewOptionListContent: TypeAlias = "OptionListContent | None | RenderableType"
-"""The type of a new item of option list content to be added to an option list.
-
-This type represents all of the types that will be accepted when adding new
-content to the option list. This is a superset of [`OptionListContent`][textual.types.OptionListContent].
-"""
+    def clear(self) -> None:
+        self.lines.clear()
+        self.heights.clear()
+        self.index_to_line.clear()
 
 
 class OptionList(ScrollView, can_focus=True):
-    """A vertical option list with bounce-bar highlighting."""
-
     ALLOW_SELECT = False
     BINDINGS: ClassVar[list[BindingType]] = [
         Binding("down", "cursor_down", "Down", show=False),
@@ -200,10 +180,13 @@ class OptionList(ScrollView, can_focus=True):
     highlighted: reactive[int | None] = reactive(None)
     """The index of the currently-highlighted option, or `None` if no option is highlighted."""
 
+    hover_option_index: reactive[int | None] = reactive(None)
+    """The index of the option under the mouse or `None`."""
+
     class OptionMessage(Message):
         """Base class for all option messages."""
 
-        def __init__(self, option_list: OptionList, index: int) -> None:
+        def __init__(self, option_list: OptionList, option: Option, index: int) -> None:
             """Initialise the option message.
 
             Args:
@@ -213,9 +196,9 @@ class OptionList(ScrollView, can_focus=True):
             super().__init__()
             self.option_list: OptionList = option_list
             """The option list that sent the message."""
-            self.option: Option = option_list.get_option_at_index(index)
+            self.option: Option = option
             """The highlighted option."""
-            self.option_id: str | None = self.option.id
+            self.option_id: str | None = option.id
             """The ID of the option that the message relates to."""
             self.option_index: int = index
             """The index of the option that the message relates to."""
@@ -230,10 +213,13 @@ class OptionList(ScrollView, can_focus=True):
             return self.option_list
 
         def __rich_repr__(self) -> rich.repr.Result:
-            yield "option_list", self.option_list
-            yield "option", self.option
-            yield "option_id", self.option_id
-            yield "option_index", self.option_index
+            try:
+                yield "option_list", self.option_list
+                yield "option", self.option
+                yield "option_id", self.option_id
+                yield "option_index", self.option_index
+            except AttributeError:
+                return
 
     class OptionHighlighted(OptionMessage):
         """Message sent when an option is highlighted.
@@ -251,442 +237,47 @@ class OptionList(ScrollView, can_focus=True):
 
     def __init__(
         self,
-        *content: NewOptionListContent,
+        *content: OptionListContent,
         name: str | None = None,
         id: str | None = None,
         classes: str | None = None,
         disabled: bool = False,
         wrap: bool = True,
-        tooltip: RenderableType | None = None,
+        markup: bool = True,
+        tooltip: VisualType | None = None,
     ):
-        """Initialise the option list.
-
-        Args:
-            *content: The content for the option list.
-            name: The name of the option list.
-            id: The ID of the option list in the DOM.
-            classes: The CSS classes of the option list.
-            disabled: Whether the option list is disabled or not.
-            wrap: Should prompts be auto-wrapped?
-            tooltip: Optional tooltip.
-        """
         super().__init__(name=name, id=id, classes=classes, disabled=disabled)
-
+        self._option_id = 0
         self._wrap = wrap
-        """Should we auto-wrap options?
+        self._markup = markup
+        self._options: list[Option] = []
+        self._id_to_option: dict[str, Option] = {}
+        self._option_to_index: dict[Option, int] = {}
 
-        If `False` options wider than the list will be truncated.
-        """
+        self._visuals: dict[int, Visual] = {}
+        self._option_render_cache: LRUCache[tuple[Option, Style], list[Strip]]
+        self._option_render_cache = LRUCache(maxsize=1024)
 
-        self._contents: list[OptionListContent] = [
-            self._make_content(item) for item in content
-        ]
-        """A list of the content of the option list.
-
-        This is *every* item that makes up the content of the option list;
-        this includes both the options *and* the separators (and any other
-        decoration we could end up adding -- although I don't anticipate
-        anything else at the moment; but padding around separators could be
-        a thing, perhaps).
-        """
-
-        self._options: list[Option] = [
-            content for content in self._contents if isinstance(content, Option)
-        ]
-        """A list of the options within the option list.
-
-        This is a list of references to just the options alone, ignoring the
-        separators and potentially any other line-oriented option list
-        content that isn't an option.
-        """
-
-        self._option_ids: dict[str, int] = {
-            option.id: index
-            for index, option in enumerate(self._options)
-            if option.id is not None
-        }
-        """A dictionary of option IDs and the option indexes they relate to."""
-
-        self._content_render_cache: LRUCache[tuple[int, str, int], list[Strip]]
-        self._content_render_cache = LRUCache(1024)
-
-        self._lines: list[tuple[int, int]] | None = None
-        self._spans: list[OptionLineSpan] | None = None
-
-        self._mouse_hovering_over: int | None = None
-        """Used to track what the mouse is hovering over."""
+        self._line_cache = _LineCache()
 
         if tooltip is not None:
             self.tooltip = tooltip
 
-        if self._options:
-            self.action_first()
+        self.add_options(content)
 
-    def _left_gutter_width(self) -> int:
-        """Returns the size of any left gutter that should be taken into account.
+    @property
+    def options(self) -> Sequence[Option]:
+        """Sequence of options in the OptionList.
 
-        Returns:
-            The width of the left gutter.
+        !!! note "This is read-only"
+
         """
-        return 0
-
-    def _on_mount(self):
-        self._populate()
-
-    def _refresh_lines(self) -> None:
-        self._lines = None
-        self._spans = None
-        self._content_render_cache.clear()
-        self._populate()
-
-    def notify_style_update(self) -> None:
-        self._content_render_cache.clear()
-        super().notify_style_update()
-
-    def _on_resize(self):
-        self._refresh_lines()
-
-    def _add_lines(
-        self, new_content: list[OptionListContent], width: int, option_index=0
-    ) -> None:
-        """Add new lines.
-
-        Args:
-            new_content: New content to add.
-            width: Width to render content.
-            option_index: Starting option index.
-        """
-        assert self._lines is not None
-        assert self._spans is not None
-
-        for index, content in enumerate(new_content, len(self._lines)):
-            if isinstance(content, Option):
-                height = len(
-                    self._render_option_content(
-                        index, content, "", width - self._left_gutter_width()
-                    )
-                )
-
-                self._spans.append(OptionLineSpan(len(self._lines), height))
-                self._lines.extend(
-                    (option_index, y_offset) for y_offset in range(height)
-                )
-                option_index += 1
-            else:
-                self._lines.append(OptionLineSpan(-1, 0))
-
-        self.virtual_size = Size(width, len(self._lines))
-        self.refresh(layout=self.styles.auto_dimensions)
-        self._scroll_update(self.virtual_size)
-
-    def _populate(self) -> None:
-        """Populate the lines data-structure."""
-
-        self._lines = []
-        self._spans = []
-
-        self._add_lines(
-            self._contents,
-            self.scrollable_content_region.width - self._left_gutter_width(),
-        )
-
-    def get_content_width(self, container: Size, viewport: Size) -> int:
-        """Get maximum width of options."""
-        console = self.app.console
-        options = console.options
-        padding = self.get_component_styles("option-list--option").padding
-        padding_width = padding.width
-        return (
-            max(
-                Measurement.get(console, options, option.prompt).maximum
-                for option in self._options
-            )
-            + padding_width
-        )
-
-    def get_content_height(self, container: Size, viewport: Size, width: int) -> int:
-        # Get the content height without requiring a refresh
-        # TODO: Internal data structure could be simplified
-        _render_option_content = self._render_option_content
-        heights = [
-            len(_render_option_content(index, option, "", width))
-            for index, option in enumerate(self._options)
-        ]
-        separator_count = sum(
-            1 for content in self._contents if isinstance(content, Separator)
-        )
-        return sum(heights) + separator_count
-
-    def _on_mouse_move(self, event: events.MouseMove) -> None:
-        """React to the mouse moving.
-
-        Args:
-            event: The mouse movement event.
-        """
-        self._mouse_hovering_over = event.style.meta.get("option")
-        self.refresh()
-
-    def _on_leave(self, _: events.Leave) -> None:
-        """React to the mouse leaving the widget."""
-        self._mouse_hovering_over = None
-
-    async def _on_click(self, event: events.Click) -> None:
-        """React to the mouse being clicked on an item.
-
-        Args:
-            event: The click event.
-        """
-        clicked_option: int | None = event.style.meta.get("option")
-        if (
-            clicked_option is not None
-            and clicked_option >= 0
-            and not self._options[clicked_option].disabled
-        ):
-            self.highlighted = clicked_option
-            self.action_select()
-
-    def _make_content(self, content: NewOptionListContent) -> OptionListContent:
-        """Convert a single item of content for the list into a content type.
-
-        Args:
-            content: The content to turn into a full option list type.
-
-        Returns:
-            The content, usable in the option list.
-        """
-        if isinstance(content, (Option, Separator)):
-            return content
-        if content is None:
-            return Separator()
-        return Option(content)
-
-    def _render_option_content(
-        self, option_index: int, content: RenderResult, component_class: str, width: int
-    ) -> list[Strip]:
-        """Render content for option and style.
-
-        Args:
-            option_index: Option index to render.
-            content: Render result for prompt.
-            component class: Additional component class.
-            width: Desired width of render.
-
-        Returns:
-            A list of strips.
-        """
-        cache_key = (option_index, component_class, width)
-        if (strips := self._content_render_cache.get(cache_key, None)) is not None:
-            return strips
-
-        visual = visualize(self, content)
-        padding = self.get_component_styles("option-list--option").padding
-        if padding:
-            visual = Padding(visual, padding)
-
-        component_class_list = ["option-list--option"]
-        if component_class:
-            component_class_list.append(component_class)
-
-        visual_style = self.get_visual_style(*component_class_list)
-
-        strips = Visual.to_strips(self, visual, width, None, visual_style, pad=True)
-        style_meta = Style.from_meta({"option": option_index})
-        strips = [strip.apply_style(style_meta) for strip in strips]
-
-        self._content_render_cache[cache_key] = strips
-        return strips
-
-    def _duplicate_id_check(self, candidate_items: list[OptionListContent]) -> None:
-        """Check the items to be added for any duplicates.
-
-        Args:
-            candidate_items: The items that are going be added.
-
-        Raises:
-            DuplicateID: If there is an attempt to use a duplicate ID.
-        """
-        # We're only interested in options, and only those that have IDs.
-        new_options = [
-            item
-            for item in candidate_items
-            if isinstance(item, Option) and item.id is not None
-        ]
-        # Get the set of new IDs that we're being given.
-        new_option_ids = {option.id for option in new_options}
-        # Now check for duplicates, both internally amongst the new items
-        # incoming, and also against all the current known IDs.
-        if len(new_options) != len(new_option_ids) or not new_option_ids.isdisjoint(
-            self._option_ids
-        ):
-            raise DuplicateID("Attempt made to add options with duplicate IDs.")
-
-    def add_options(self, items: Iterable[NewOptionListContent]) -> Self:
-        """Add new options to the end of the option list.
-
-        Args:
-            items: The new items to add.
-
-        Returns:
-            The `OptionList` instance.
-
-        Raises:
-            DuplicateID: If there is an attempt to use a duplicate ID.
-
-        Note:
-            All options are checked for duplicate IDs *before* any option is
-            added. A duplicate ID will cause none of the passed items to be
-            added to the option list.
-        """
-        # Only work if we have items to add; but don't make a fuss out of
-        # zero items to add, just carry on like nothing happened.
-        if self._lines is None:
-            self._lines = []
-        if self._spans is None:
-            self._spans = []
-        new_items = list(items)
-        if new_items:
-            option_index = len(self._options)
-            # Turn any incoming values into valid content for the list.
-            content = [self._make_content(item) for item in new_items]
-            self._duplicate_id_check(content)
-            self._contents.extend(content)
-            # Pull out the content that is genuine options, create any new
-            # ID mappings required, then add the new options to the option
-            # list.
-            new_options = [item for item in content if isinstance(item, Option)]
-            for new_option_index, new_option in enumerate(
-                new_options, start=len(self._options)
-            ):
-                if new_option.id:
-                    self._option_ids[new_option.id] = new_option_index
-            self._options.extend(new_options)
-
-            self._add_lines(
-                content,
-                self.scrollable_content_region.width - self._left_gutter_width(),
-                option_index=option_index,
-            )
-            self.refresh(layout=True)
-        return self
-
-    def add_option(self, item: NewOptionListContent = None) -> Self:
-        """Add a new option to the end of the option list.
-
-        Args:
-            item: The new item to add.
-
-        Returns:
-            The `OptionList` instance.
-
-        Raises:
-            DuplicateID: If there is an attempt to use a duplicate ID.
-        """
-        return self.add_options([item])
-
-    def _remove_option(self, index: int) -> None:
-        """Remove an option from the option list.
-
-        Args:
-            index: The index of the item to remove.
-
-        Raises:
-            IndexError: If there is no option of the given index.
-        """
-        option = self._options[index]
-        del self._options[index]
-        del self._contents[self._contents.index(option)]
-        # Decrement index of options after the one we just removed.
-        self._option_ids = {
-            option_id: option_index - 1 if option_index > index else option_index
-            for option_id, option_index in self._option_ids.items()
-            if option_index != index
-        }
-        self._refresh_lines()
-        # Force a re-validation of the highlight.
-        self.highlighted = self.highlighted
-        self._mouse_hovering_over = None
-
-    def remove_option(self, option_id: str) -> Self:
-        """Remove the option with the given ID.
-
-        Args:
-            option_id: The ID of the option to remove.
-
-        Returns:
-            The `OptionList` instance.
-
-        Raises:
-            OptionDoesNotExist: If no option has the given ID.
-        """
-        self._remove_option(self.get_option_index(option_id))
-        return self
-
-    def remove_option_at_index(self, index: int) -> Self:
-        """Remove the option at the given index.
-
-        Args:
-            index: The index of the option to remove.
-
-        Returns:
-            The `OptionList` instance.
-
-        Raises:
-            OptionDoesNotExist: If there is no option with the given index.
-        """
-        try:
-            self._remove_option(index)
-        except IndexError:
-            raise OptionDoesNotExist(
-                f"There is no option with an index of {index!r}"
-            ) from None
-        return self
-
-    def _replace_option_prompt(self, index: int, prompt: RenderableType) -> None:
-        """Replace the prompt of an option in the list.
-
-        Args:
-            index: The index of the option to replace the prompt of.
-            prompt: The new prompt for the option.
-
-        Raises:
-            OptionDoesNotExist: If there is no option with the given index.
-        """
-        self.get_option_at_index(index).set_prompt(prompt)
-        self._refresh_lines()
-
-    def replace_option_prompt(self, option_id: str, prompt: RenderableType) -> Self:
-        """Replace the prompt of the option with the given ID.
-
-        Args:
-            option_id: The ID of the option to replace the prompt of.
-            prompt: The new prompt for the option.
-
-        Returns:
-            The `OptionList` instance.
-
-        Raises:
-            OptionDoesNotExist: If no option has the given ID.
-        """
-        self._replace_option_prompt(self.get_option_index(option_id), prompt)
-        return self
-
-    def replace_option_prompt_at_index(
-        self, index: int, prompt: RenderableType
-    ) -> Self:
-        """Replace the prompt of the option at the given index.
-
-        Args:
-            index: The index of the option to replace the prompt of.
-            prompt: The new prompt for the option.
-
-        Returns:
-            The `OptionList` instance.
-
-        Raises:
-            OptionDoesNotExist: If there is no option with the given index.
-        """
-        self._replace_option_prompt(index, prompt)
-        return self
+        return self._options
+
+    @property
+    def option_count(self) -> int:
+        """The number of options."""
+        return len(self._options)
 
     def clear_options(self) -> Self:
         """Clear the content of the option list.
@@ -694,13 +285,107 @@ class OptionList(ScrollView, can_focus=True):
         Returns:
             The `OptionList` instance.
         """
-        self._contents.clear()
-        self._options.clear()
-        self._option_ids.clear()
+        self._line_cache.clear()
+        self._option_render_cache.clear()
         self.highlighted = None
-        self._mouse_hovering_over = None
-        self._refresh_lines()
+        self.refresh()
         return self
+
+    def add_options(self, new_options: Iterable[OptionListContent]) -> Self:
+        """Add new options.
+
+        Args:
+            new_options: Content of new options.
+        """
+
+        options = self._options
+        add_option = self._options.append
+        for prompt in new_options:
+            if isinstance(prompt, Option):
+                option = prompt
+            elif prompt is None:
+                if options:
+                    options[-1]._divider = True
+                continue
+            else:
+                option = Option(prompt)
+            hash(option)
+            self._option_to_index[option] = len(options)
+            add_option(option)
+            if option._id is not None:
+                if option._id in self._id_to_option:
+                    raise DuplicateID("Unable to add {option!r} due to duplicate ID")
+                self._id_to_option[option._id] = option
+
+        return self
+
+    def add_option(self, option: Option | VisualType | None) -> Self:
+        """Add a new option to the end of the option list.
+
+        Args:
+            option: New option to add, or `None` for a separator.
+
+        Returns:
+            The `OptionList` instance.
+
+        Raises:
+            DuplicateID: If there is an attempt to use a duplicate ID.
+        """
+        self.add_options([option])
+        return self
+
+    def get_option(self, option_id: str) -> Option:
+        """Get the option with the given ID.
+
+        Args:
+            option_id: The ID of the option to get.
+
+        Returns:
+            The option with the ID.
+
+        Raises:
+            OptionDoesNotExist: If no option has the given ID.
+        """
+        try:
+            return self._id_to_option[option_id]
+        except KeyError:
+            raise OptionDoesNotExist(
+                f"There is no option with an ID of {option_id!r}"
+            ) from None
+
+    def get_option_index(self, option_id: str) -> int:
+        """Get the index (offset in `self.options`) of the option with the given ID.
+
+        Args:
+            option_id: The ID of the option to get the index of.
+
+        Returns:
+            The index of the item with the given ID.
+
+        Raises:
+            OptionDoesNotExist: If no option has the given ID.
+        """
+        option = self.get_option(option_id)
+        return self._option_to_index[option]
+
+    def get_option_at_index(self, index: int) -> Option:
+        """Get the option at the given index.
+
+        Args:
+            index: The index of the option to get.
+
+        Returns:
+            The option at that index.
+
+        Raises:
+            OptionDoesNotExist: If there is no option with the given index.
+        """
+        try:
+            return self._options[index]
+        except IndexError:
+            raise OptionDoesNotExist(
+                f"There is no option with an index of {index}"
+            ) from None
 
     def _set_option_disabled(self, index: int, disabled: bool) -> Self:
         """Set the disabled state of an option in the list.
@@ -712,7 +397,7 @@ class OptionList(ScrollView, can_focus=True):
         Returns:
             The `OptionList` instance.
         """
-        self._options[index].disabled = disabled
+        self._options[index]._disabled = disabled
         if index == self.highlighted:
             self.highlighted = _widget_navigation.find_next_enabled(
                 self._options, anchor=index, direction=1
@@ -781,134 +466,320 @@ class OptionList(ScrollView, can_focus=True):
         """
         return self.disable_option_at_index(self.get_option_index(option_id))
 
-    @property
-    def option_count(self) -> int:
-        """The count of options."""
-        return len(self._options)
-
-    def get_option_at_index(self, index: int) -> Option:
-        """Get the option at the given index.
+    def _remove_option(self, option: Option) -> Self:
+        """Remove the option with the given ID.
 
         Args:
-            index: The index of the option to get.
+            option: The Option to return.
 
         Returns:
-            The option at that index.
+            The `OptionList` instance.
+
+        Raises:
+            OptionDoesNotExist: If no option has the given ID.
+        """
+
+        index = self._option_to_index[option]
+        for option in self.options[index + 1 :]:
+            current_index = self._option_to_index[option]
+            self._option_to_index[option] = current_index - 1
+
+        option = self._options[index]
+        del self._options[index]
+        if option._id is not None:
+            del self._id_to_option[option._id]
+        del self._option_to_index[option]
+        self.refresh()
+        return self
+
+    def remove_option(self, option_id: str) -> Self:
+        """Remove the option with the given ID.
+
+        Args:
+            option_id: The ID of the option to remove.
+
+        Returns:
+            The `OptionList` instance.
+
+        Raises:
+            OptionDoesNotExist: If no option has the given ID.
+        """
+        option = self.get_option(option_id)
+        return self._remove_option(option)
+
+    def remove_option_at_index(self, index: int) -> Self:
+        """Remove the option at the given index.
+
+        Args:
+            index: The index of the option to remove.
+
+        Returns:
+            The `OptionList` instance.
 
         Raises:
             OptionDoesNotExist: If there is no option with the given index.
         """
-        try:
-            return self._options[index]
-        except IndexError:
-            raise OptionDoesNotExist(
-                f"There is no option with an index of {index}"
-            ) from None
+        option = self._options[index]
+        return self._remove_option(option)
 
-    def get_option(self, option_id: str) -> Option:
-        """Get the option with the given ID.
+    @property
+    def _lines(self) -> Sequence[tuple[int, int]]:
+        """A sequence of pairs of ints for each line, used internally.
 
-        Args:
-            option_id: The ID of the option to get.
+        The first int is the index of the option, and second is the line offset.
+
+        !!! note "This is read-only"
 
         Returns:
-            The option with the ID.
-
-        Raises:
-            OptionDoesNotExist: If no option has the given ID.
+            A sequence of tuples.
         """
-        return self.get_option_at_index(self.get_option_index(option_id))
+        self._update_lines()
+        return self._line_cache.lines
 
-    def get_option_index(self, option_id: str) -> int:
-        """Get the index of the option with the given ID.
+    @property
+    def _heights(self) -> dict[int, int]:
+        self._update_lines()
+        return self._line_cache.heights
+
+    @property
+    def _index_to_line(self) -> dict[int, int]:
+        self._update_lines()
+        return self._line_cache.index_to_line
+
+    def _clear_caches(self) -> None:
+        self._option_render_cache.clear()
+        self._line_cache.clear()
+
+    def notify_style_update(self) -> None:
+        self._clear_caches()
+
+    def _on_resize(self):
+        self._clear_caches()
+        self.refresh()
+
+    def on_show(self) -> None:
+        self.scroll_to_highlight()
+
+    async def _on_click(self, event: events.Click) -> None:
+        """React to the mouse being clicked on an item.
 
         Args:
-            option_id: The ID of the option to get the index of.
+            event: The click event.
+        """
+        clicked_option: int | None = event.style.meta.get("option")
+        if (
+            clicked_option is not None
+            and clicked_option >= 0
+            and not self._options[clicked_option].disabled
+        ):
+            self.highlighted = clicked_option
+            self.action_select()
+
+    def _get_left_gutter_width(self) -> int:
+        """Returns the size of any left gutter that should be taken into account.
 
         Returns:
-            The index of the item with the given ID.
-
-        Raises:
-            OptionDoesNotExist: If no option has the given ID.
+            The width of the left gutter.
         """
-        try:
-            return self._option_ids[option_id]
-        except KeyError:
-            raise OptionDoesNotExist(
-                f"There is no option with an ID of '{option_id}'"
-            ) from None
+        return 0
+
+    def _on_mouse_move(self, event: events.MouseMove) -> None:
+        """React to the mouse moving.
+
+        Args:
+            event: The mouse movement event.
+        """
+        self.hover_option_index = event.style.meta.get("option")
+
+    def _on_leave(self, _: events.Leave) -> None:
+        """React to the mouse leaving the widget."""
+        self._mouse_hovering_over = None
+
+    def _get_visual(self, option: Option) -> Visual:
+        """Get a visual for the given option.
+
+        Args:
+            option: An option.
+
+        Returns:
+            A Visual.
+
+        """
+        if (visual := option._visual) is None:
+            visual = visualize(self, option.prompt, markup=self._markup)
+            option._visual = visual
+        return visual
+
+    def _get_visual_from_index(self, index: int) -> Visual:
+        """Get a visual from the given index.
+
+        Args:
+            index: An index (offset in self.options).
+
+        Returns:
+            A Visual.
+        """
+        option = self.get_option_at_index(index)
+        return self._get_visual(option)
+
+    def _get_option_render(self, option: Option, style: Style) -> list[Strip]:
+        """Get rendered option with a given style.
+
+        Args:
+            style: Style of render.
+            index: Index of the option.
+
+        Returns:
+            A list of strips.
+        """
+        padding = self.get_component_styles("option-list--option").padding
+        width = self.content_region.width - self._get_left_gutter_width()
+        cache_key = (option, style)
+        if (strips := self._option_render_cache.get(cache_key)) is None:
+            visual = self._get_visual(option)
+            if padding:
+                visual = Padding(visual, padding)
+            index = self._option_to_index[option]
+            strips = visual.to_strips(self, visual, width, None, style)
+            strips = [
+                strip.extend_cell_length(width, style.rich_style).apply_meta(
+                    {"option": index}
+                )
+                for strip in strips
+            ]
+            if option._divider:
+                style = self.get_visual_style("option-list--separator")
+                rule_segments = [Segment("─" * width, style.rich_style)]
+                strips.append(Strip(rule_segments, width))
+            self._option_render_cache[cache_key] = strips
+        return strips
+
+    def _update_lines(self) -> None:
+        line_cache = self._line_cache
+        lines = line_cache.lines
+        last_index = lines[-1][0] if lines else 0
+        get_visual = self._get_visual
+        width = self.scrollable_content_region.width
+
+        if last_index < len(self.options) - 1:
+            styles = self.get_component_styles("option-list--option")
+
+            for index, option in enumerate(self.options[last_index:], last_index):
+                line_cache.index_to_line[index] = len(line_cache.lines)
+                line_count = (
+                    get_visual(option).get_height(styles, width) + option._divider
+                )
+                line_cache.heights[index] = line_count
+                line_cache.lines.extend(
+                    [(index, line_no) for line_no in range(0, line_count)]
+                )
+
+        last_divider = self.options and self.options[-1]._divider
+        self.virtual_size = Size(
+            self.content_region.width, len(lines) - (1 if last_divider else 0)
+        )
+
+    def get_content_width(self, container: Size, viewport: Size) -> int:
+        """Get maximum width of options."""
+
+        styles = self.styles
+        get_visual_from_index = self._get_visual_from_index
+        padding = self.get_component_styles("option-list--option").padding
+        gutter_width = self._get_left_gutter_width()
+        container_width = container.width
+        width = (
+            max(
+                get_visual_from_index(index).get_optimal_width(styles, container_width)
+                for index in range(len(self.options))
+            )
+            + padding.width
+            + gutter_width
+        )
+        return width
+
+    def get_content_height(self, container: Size, viewport: Size, width: int) -> int:
+        """Get height for the given width."""
+        styles = cast(RulesMap, self.get_component_styles("option-list--option"))
+        get_visual = self._get_visual
+        height = sum(
+            (
+                get_visual(option).get_height(styles, width)
+                + (1 if option._divider and not last else 0)
+            )
+            for last, option in loop_last(self.options)
+        )
+        return height
+
+    def _get_line(self, style: Style, y: int) -> Strip:
+        index, line_offset = self._lines[y]
+        option = self.get_option_at_index(index)
+        strips = self._get_option_render(option, style)
+        return strips[line_offset]
 
     def render_line(self, y: int) -> Strip:
-        assert self._lines is not None
-        if not self._lines:
-            self._populate()
-
-        _scroll_x, scroll_y = self.scroll_offset
-        line_number = scroll_y + y
-
+        line_number = self.scroll_offset.y + y
         try:
-            option_index, y_offset = self._lines[line_number]
+            option_index, line_offset = self._lines[line_number]
         except IndexError:
-            return Strip([])
+            return Strip.blank(self.scrollable_content_region.width)
+        option = self.options[option_index]
+        mouse_over = self.hover_option_index == option_index
+        component_classes: tuple[str, ...] = ("option-list--option",)
+        if option.disabled:
+            component_classes = (
+                "option-list--option",
+                "option-list--option-disabled",
+            )
+        elif self.highlighted == option_index:
+            component_classes = (
+                "option-list--option",
+                "option-list--option-highlighted",
+            )
+        elif mouse_over:
+            component_classes = (
+                "option-list--option",
+                "option-list--option-hover",
+            )
 
-        renderable = (
-            Rule(style=self.get_component_rich_style("option-list--separator"))
-            if option_index == -1
-            else self._options[option_index]
-        )
-
-        mouse_over = self._mouse_hovering_over == option_index
-
-        component_class: str = ""
-
-        if option_index == -1:
-            component_class = "option-list--separator"
-        else:
-            try:
-                option = self._options[option_index]
-            except IndexError:
-                pass
-            else:
-                if option.disabled:
-                    component_class = "option-list--option-disabled"
-                elif self.highlighted == option_index:
-                    component_class = "option-list--option-highlighted"
-                elif mouse_over:
-                    component_class = "option-list--option-hover"
-
-        strips = self._render_option_content(
-            option_index,
-            renderable,
-            component_class,
-            self.scrollable_content_region.width - self._left_gutter_width(),
-        )
-        try:
-            strip = strips[y_offset]
-        except IndexError:
-            return Strip([])
+        style = self.get_visual_style(*component_classes)
+        strips = self._get_option_render(option, style)
+        strip = strips[line_offset]
         return strip
 
+    def validate_highlighted(self, highlighted: int | None) -> int | None:
+        """Validate the `highlighted` property value on access."""
+        if highlighted is None or not self.options:
+            return None
+        elif highlighted < 0:
+            return 0
+        elif highlighted >= len(self.options):
+            return len(self.options) - 1
+        return highlighted
+
+    def watch_highlighted(self, highlighted: int | None) -> None:
+        """React to the highlighted option having changed."""
+        if highlighted is None:
+            return
+
+        if not self._options[highlighted].disabled:
+            self.scroll_to_highlight()
+
+        option: Option
+        if highlighted is None:
+            option = None
+        else:
+            option = self.options[highlighted]
+        self.post_message(self.OptionHighlighted(self, option, highlighted))
+
     def scroll_to_highlight(self, top: bool = False) -> None:
-        """Ensure that the highlighted option is in view.
-
-        Args:
-            top: Scroll highlight to top of the list.
-        """
         highlighted = self.highlighted
-
         if highlighted is None or not self.is_mounted:
             return
 
-        if not self._spans:
-            self._populate()
+        y = self._index_to_line[highlighted]
+        option = self.options[highlighted]
+        height = self._heights[highlighted] - option._divider
 
-        try:
-            y, height = self._spans[highlighted]
-        except IndexError:
-            # Index error means we're being asked to scroll to a highlight
-            # before all the tracking information has been worked out.
-            # That's fine; let's just NoP that.
-            return
         self.scroll_to_region(
             Region(0, y, self.scrollable_content_region.width, height),
             force=True,
@@ -917,31 +788,10 @@ class OptionList(ScrollView, can_focus=True):
             immediate=True,
         )
 
-    def on_show(self) -> None:
-        if self.highlighted is not None:
-            self.scroll_to_highlight()
-
-    def validate_highlighted(self, highlighted: int | None) -> int | None:
-        """Validate the `highlighted` property value on access."""
-        if highlighted is None or not self._options:
-            return None
-        elif highlighted < 0:
-            return 0
-        elif highlighted >= len(self._options):
-            return len(self._options) - 1
-
-        return highlighted
-
-    def watch_highlighted(self, highlighted: int | None) -> None:
-        """React to the highlighted option having changed."""
-        if highlighted is not None and not self._options[highlighted].disabled:
-            self.scroll_to_highlight()
-            self.post_message(self.OptionHighlighted(self, highlighted))
-
     def action_cursor_up(self) -> None:
         """Move the highlight up to the previous enabled option."""
         self.highlighted = _widget_navigation.find_next_enabled(
-            self._options,
+            self.options,
             anchor=self.highlighted,
             direction=-1,
         )
@@ -949,106 +799,89 @@ class OptionList(ScrollView, can_focus=True):
     def action_cursor_down(self) -> None:
         """Move the highlight down to the next enabled option."""
         self.highlighted = _widget_navigation.find_next_enabled(
-            self._options,
+            self.options,
             anchor=self.highlighted,
             direction=1,
         )
 
     def action_first(self) -> None:
         """Move the highlight to the first enabled option."""
-        self.highlighted = _widget_navigation.find_first_enabled(self._options)
+        self.highlighted = _widget_navigation.find_first_enabled(self.options)
 
     def action_last(self) -> None:
         """Move the highlight to the last enabled option."""
-        self.highlighted = _widget_navigation.find_last_enabled(self._options)
+        self.highlighted = _widget_navigation.find_last_enabled(self.options)
 
-    def _page(self, direction: Direction) -> None:
-        """Move the highlight roughly by one page in the given direction.
+    def _move_page(self, direction: _widget_navigation.Direction) -> None:
+        """Move the height roughly by one page in the given direction.
 
-        The highlight will tentatively move by exactly one page.
-        If this would result in highlighting a disabled option, instead we look for
-        an enabled option "further down" the list of options.
-        If there are no such enabled options, we fallback to the "last" enabled option.
-        (The meaning of "further down" and "last" depend on the direction specified.)
+        This method will attempt to avoid selecting a disabled option.
 
         Args:
-            direction: The direction to head, -1 for up and 1 for down.
+            direction: `-1` to move up a page, `1` to move down a page.
         """
 
-        # If we find ourselves in a position where we don't know where we're
-        # going, we need a fallback location. Where we go will depend on the
-        # direction.
-        assert self._spans is not None
-        assert self._lines is not None
+        height = self.content_region.height
 
-        fallback = self.action_first if direction == -1 else self.action_last
+        option_index = self.highlighted or 0
 
-        highlighted = self.highlighted
-        if highlighted is None:
-            # There is no highlight yet so let's go to the default position.
-            fallback()
-        else:
-            # We want to page roughly by lines, but we're dealing with
-            # options that can be a varying number of lines in height. So
-            # let's start with the target line alone.
-            target_line = max(
-                0,
-                self._spans[highlighted].first
-                + (direction * self.scrollable_content_region.height),
-            )
-            try:
-                # Now that we've got a target line, let's figure out the
-                # index of the target option.
-                target_option: int | None = self._lines[target_line][0]
-            except IndexError:
-                # An index error suggests we've gone out of bounds, let's
-                # settle on whatever the call thinks is a good place to wrap
-                # to.
-                fallback()
-            else:
-                # Looks like we've figured where we'd like to jump to, we
-                # just need to make sure we jump to an option that's enabled.
-                if target_option is not None:
-                    target_option = _widget_navigation.find_next_enabled_no_wrap(
-                        candidates=self._options,
-                        anchor=target_option,
-                        direction=direction,
-                        with_anchor=True,
-                    )
-                    # If we couldn't find an enabled option that's at least one page
-                    # away from the current one, we instead move less than one page
-                    # to the last enabled option in the correct direction.
-                    if target_option is None:
-                        fallback()
-                    else:
-                        self.highlighted = target_option
+        y = min(
+            self._index_to_line[option_index] + direction * height,
+            len(self._lines) - 1,
+        )
+        option_index = self._lines[y][0]
 
-    def action_page_up(self) -> None:
-        """Move the highlight up roughly by one page."""
-        self._page(-1)
+        target_option = _widget_navigation.find_next_enabled_no_wrap(
+            candidates=self._options,
+            anchor=option_index,
+            direction=direction,
+            with_anchor=True,
+        )
+        if target_option is not None:
+            self.highlighted = target_option
 
-    def action_page_down(self) -> None:
-        """Move the highlight down roughly by one page."""
-        self._page(1)
+    def action_page_up(self):
+        """Move the highlight up one page."""
+        self._move_page(-1)
+
+    def action_page_down(self):
+        """Move the highlight down one page."""
+        self._move_page(1)
 
     def action_select(self) -> None:
-        """Select the currently-highlighted option.
+        """Select the currently highlighted option.
 
-        If no option is selected, then nothing happens. If an option is
-        selected, a [OptionList.OptionSelected][textual.widgets.OptionList.OptionSelected]
-        message will be posted.
+        If an option is selected then a
+        [OptionList.OptionSelected][textual.widgets.OptionList.OptionSelected] will be posted.
         """
+        if self.highlighted is None:
+            return
         highlighted = self.highlighted
-        if highlighted is not None and not self._options[highlighted].disabled:
-            self.post_message(self.OptionSelected(self, highlighted))
+        option = self._options[highlighted]
+        if highlighted is not None and not option.disabled:
+            self.post_message(self.OptionSelected(self, option, highlighted))
 
 
 if __name__ == "__main__":
     from textual.app import App, ComposeResult
 
-    class OptionApp(App):
-        def compose(self) -> ComposeResult:
-            yield OptionList("Foo", "Bar", "Baz")
+    TEXT = """I must not fear.
+Fear is the [u]mind-killer[/u].
+Fear is the little-death that brings total obliteration.
+I will face my fear.
+I will permit it to pass over me and through me.
+And when it has gone past, I will turn the inner eye to see its path.
+Where the fear has gone there will be nothing. Only I will remain."""
 
-    app = OptionApp()
+    class OLApp(App):
+
+        def compose(self) -> ComposeResult:
+            yield OptionList(
+                *(
+                    ["Hello", "World!", None, TEXT, None, "Foo", "Bar", "Baz", None]
+                    * 100
+                )
+            )
+
+    app = OLApp()
     app.run()
