+++ src/textual/widgets/option_list.py
@@ -1,8 +1,3 @@
-from textual.widgets._option_list import (
-    DuplicateID,
-    Option,
-    OptionDoesNotExist,
-    Separator,
-)
+from textual.widgets._option_list import DuplicateID, Option, OptionDoesNotExist
 
-__all__ = ["DuplicateID", "Option", "OptionDoesNotExist", "Separator"]
+__all__ = ["DuplicateID", "Option", "OptionDoesNotExist"]
